backend/requirements.txt
# FastAPI and web server
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0

# Vector database and embeddings
qdrant-client>=1.3.0
sentence-transformers>=2.2.2
backoff>=2.2.1  # For retry logic in upload_vectors.py

# AI and LLM integration
openai>=1.75.0
tenacity>=9.1.2
tiktoken>=0.5.0  # For token counting in parallel_processor.py

# Data processing and documents
pandas>=2.0.0
pyarrow>=12.0.0
fastparquet>=2023.4.0
reportlab>=4.0.0
tqdm>=4.65.0  # For progress bars

# SQLite for database
# (sqlite3 is included in Python standard library)

# Hugging Face for dataset access
huggingface-hub>=0.16.4  # For downloading from HF datasets

# Machine learning
numpy>=1.24.0
torch>=2.0.0  # Required by sentence-transformers
transformers>=4.30.0  # Required by sentence-transformers

# Utilities
multiprocessing-logging>=0.3.4  # Better logging for parallel processing

# Testing
pytest>=8.3.5
pytest-asyncio>=0.26.0

# Development tools (optional)
ipython>=8.12.0
jupyter>=1.0.0  # If using notebooks for exploration